requirements.txt
Flask==2.1.0
gunicorn==20.1.0
selenium==3.141.0
webbot
python-dotenv
firebase-admin